Extract Keywords and Implicits objects
  ref #22
