+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/package.scala
@@ -4,24 +4,6 @@ import diplomacy.board.{ Power, Location, DiplomacyUnit }
 import diplomacy.board.Power
 
 package object rule {
-  object Keywords {
-    /*  Military Branch */
-    val Army = MilitaryBranch.Army
-    val Fleet = MilitaryBranch.Fleet
-
-    /* Phase */
-    val Movement = Phase.Movement
-    val Retreat = Phase.Retreat
-    val Build = Phase.Build
-  }
-
-  object Implicits {
-    /* State */
-    implicit class Turn2State[Turn_ <: Turn](turn: Turn_) {
-      def -(phase: Phase): State[Turn_] = State(turn, phase)
-    }
-  }
-
   /* Type Aliases */
   type Location[Power_ <: Power] = diplomacy.board.Location[Power_, MilitaryBranch]
   type DiplomacyUnit[Power_ <: Power] =
